fix: resolve all mypy type errors
- config.py: add type: ignore[import-untyped] for pyyaml
- network.py: explicit str() cast on getaddrinfo result
- memory/store.py: annotate `out` list and declare `_backend` union type
- tools/filesystem.py: widen _fmt_size param from int to float
- tools/voice_tools.py: use TYPE_CHECKING guard for VoiceInterface import

+++ llmos/tools/filesystem.py
@@ -10,7 +10,7 @@ from pathlib import Path
 from .registry import tool
 
 
-def _fmt_size(n: int) -> str:
+def _fmt_size(n: float) -> str:
     for unit in ("B", "K", "M", "G", "T"):
         if n < 1024:
             return f"{n:.0f}{unit}"
